# preprocess converting sentence into set of words