Moving set operations to outside the main loop to significantly speed up the process.

before, on 5000 sentences (MacPro M2 8GB):
```
78.91s user 0.91s system 102% cpu 1:18.14 total
MD5 (test-before/output.csv) = f798a4625a1ecbb5e09d69776dacd583
```

after, on 5000 sentences:
```
6.97s user 0.43s system 127% cpu 5.814 total
MD5 (test-after/output.csv) = f798a4625a1ecbb5e09d69776dacd583
```

after, on 79000 sentences:
```
1327.29s user 6.15s system 99% cpu 22:14.58 total
```